tidy up function system
